requirements.txt
gooddata_sdk>=1.2.0
python-dotenv>=0.20.0
